requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Scientific Computing
scipy==1.11.4
numpy==1.26.3
pandas==2.1.4

# Chat Interface
streamlit==1.30.0
streamlit-chat==0.1.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.26.0

# Data Validation
python-multipart==0.0.6
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
pypdf==3.17.4

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
torch==2.2.0

# Future: Neo4j integration
# neo4j==5.16.0

# Future: Claude integration
# anthropic==0.8.1
plotly>=5.22.0
